Addressing review comments  - include cleanup  - enum to enum class  - catch2 as submodule  - avoid usage of c cast  - select serial or parallel implementation depending on MPI and HDF5 found
Fixed bug when writing to HDF5 file on serial implementation

+++ src/reports/library/compartment_report.cpp
@@ -1,15 +1,12 @@
-#include <stdlib.h>
-#include <new>
 #include <iostream>
-#include <stdlib.h>
 
 #include "compartment_report.hpp"
 
 CompartmentReport::CompartmentReport(const std::string& report_name, double tstart, double tend, double dt)
 : Report(report_name, tstart, tend, dt) {}
 
-int CompartmentReport::get_total_compartments() {
-    int total = 0;
+size_t CompartmentReport::get_total_compartments() {
+    size_t total = 0;
     for(auto& kv: *m_cells) {
         total += kv.second.get_num_compartments();
     }
